Renames the attr_map mod to params
